refactor: extract create example app options out
