+++ projects/ngx-meta/example-apps/src/create-example-app-options.ts
@@ -1,0 +1,8 @@
+import { AngularCliVersion } from './angular/index.js'
+
+export interface CreateExampleAppOptions {
+  readonly angularCliVersion: AngularCliVersion
+  readonly baseAppDir?: string
+  readonly noCleanup?: boolean
+  readonly tmpDir?: string
+}
